no long

+++ test/test_rangeify.py
@@ -79,8 +79,9 @@ class TestRangeifyEdgeCase(unittest.TestCase):
   def test_pcontig_multi_gather(self):
     # regression test: local bufferize must have device set for const_like to work
     with Context(PCONTIG=2):
-      forest = Tensor(list(range(8)), dtype='uint')
-      idx = Tensor([0, 0], dtype='uint')
+      # NOTE: with uint type, this will become a long and fail on WEBGPU
+      forest = Tensor(list(range(8)), dtype='int')
+      idx = Tensor([0, 0], dtype='int')
       node_val = forest.gather(0, idx)
       idx2 = idx * 2 + 1
       node_val2 = forest.gather(0, idx2)
